+++ src/dropin.js
@@ -24,7 +24,21 @@ var mainHTML = fs.readFileSync(__dirname + '/html/main.html', 'utf8');
 var svgHTML = fs.readFileSync(__dirname + '/html/svgs.html', 'utf8');
 
 var ASSETS_URL = 'https://assets.braintreegateway.com';
+var PASS_THROUGH_EVENTS = [
+  'paymentMethodRequestable',
+  'noPaymentMethodRequestable',
+  'paymentOptionSelected',
 
+  // Card View Events
+  'card:binAvailable',
+  'card:blur',
+  'card:cardTypeChange',
+  'card:empty',
+  'card:focus',
+  'card:inputSubmitRequest',
+  'card:notEmpty',
+  'card:validityChange'
+];
 var UPDATABLE_CONFIGURATION_OPTIONS = [
   paymentOptionIDs.paypal,
   paymentOptionIDs.paypalCredit,
@@ -51,6 +65,7 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
  * @property {string} description A human-readable description.
  * @property {string} type The payment method type, always `CreditCard` when the method requested is a card.
  * @property {object} binData Information about the card based on the bin. Documented {@link Dropin~binData|here}.
+ * @property {?boolean} vaulted If present and true, indicates that the payment method refers to a vaulted payment method.
  * @property {?string} deviceData If data collector is configured, the device data property to be used when making a transaction.
  * @property {?boolean} liabilityShifted If 3D Secure is configured, whether or not liability did shift.
  * @property {?boolean} liabilityShiftPossible If 3D Secure is configured, whether or not liability shift is possible.
@@ -60,6 +75,7 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
 /**
  * @typedef {object} Dropin~paypalPaymentMethodPayload
  * @property {string} nonce The payment method nonce, used by your server to charge the PayPal account.
+ * @property {?boolean} vaulted If present and true, indicates that the payment method refers to a vaulted payment method.
  * @property {object} details Additional PayPal account details. See a full list of details in the [PayPal client reference](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/PayPalCheckout.html#~tokenizePayload).
  * @property {string} type The payment method type, always `PayPalAccount` when the method requested is a PayPal account.
  * @property {?string} deviceData If data collector is configured, the device data property to be used when making a transaction.
@@ -68,6 +84,7 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
 /**
  * @typedef {object} Dropin~applePayPaymentMethodPayload
  * @property {string} nonce The payment method nonce, used by your server to charge the Apple Pay provided card.
+ * @property {?boolean} vaulted If present and true, indicates that the payment method refers to a vaulted payment method.
  * @property {string} details.cardType Type of card, ex: Visa, Mastercard.
  * @property {string} details.cardHolderName The name of the card holder.
  * @property {string} details.dpanLastTwo Last two digits of card number.
@@ -87,6 +104,7 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
 /**
  * @typedef {object} Dropin~venmoPaymentMethodPayload
  * @property {string} nonce The payment method nonce, used by your server to charge the Venmo account.
+ * @property {?boolean} vaulted If present and true, indicates that the payment method refers to a vaulted payment method.
  * @property {string} details.username The Venmo username.
  * @property {string} type The payment method type, always `VenmoAccount` when the method requested is a Venmo account.
  * @property {?string} deviceData If data collector is configured, the device data property to be used when making a transaction.
@@ -95,6 +113,7 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
 /**
  * @typedef {object} Dropin~googlePayPaymentMethodPayload
  * @property {string} nonce The payment method nonce, used by your server to charge the Google Pay card.
+ * @property {?boolean} vaulted If present and true, indicates that the payment method refers to a vaulted payment method.
  * @property {string} details.cardType Type of card, ex: Visa, Mastercard.
  * @property {string} details.lastFour The last 4 digits of the card.
  * @property {string} details.lastTwo The last 2 digits of the card.
@@ -131,9 +150,20 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
  * @param {string} event The name of the event to which you are subscribing.
  * @param {function} handler A callback to handle the event.
  * @description Subscribes a handler function to a named event. `event` should be one of the following:
+ *
  *  * [`paymentMethodRequestable`](#event:paymentMethodRequestable)
  *  * [`noPaymentMethodRequestable`](#event:noPaymentMethodRequestable)
  *  * [`paymentOptionSelected`](#event:paymentOptionSelected)
+ *
+ *  _Card View Specific Events_
+ *  * [`card:binAvailable`](#event:card:binAvailable)
+ *  * [`card:blur`](#event:card:blur)
+ *  * [`card:cardTypeChange`](#event:card:cardTypeChange)
+ *  * [`card:empty`](#event:card:empty)
+ *  * [`card:focus`](#event:card:focus)
+ *  * [`card:inputSubmitRequest`](#event:card:inputSubmitRequest)
+ *  * [`card:notEmpty`](#event:card:notEmpty)
+ *  * [`card:validityChange`](#event:card:validityChange)
  * @returns {void}
  * @example
  * <caption>Dynamically enable or disable your submit button based on whether or not the payment method is requestable</caption>
@@ -206,6 +236,22 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
  *     }
  *   });
  * });
+ * @example
+ * <caption>Listen on various events from the card view</caption>
+ * braintree.dropin.create({
+ *   authorization: 'CLIENT_AUTHORIZATION',
+ *   container: '#dropin-container'
+ * }, function (err, dropinInstance) {
+ *   dropinInstance.on('card:focus', function (event) {
+ *     // a card field was focussed
+ *   });
+ *   dropinInstance.on('card:blur', function (event) {
+ *     // a card field was blurred
+ *   });
+ *   dropinInstance.on('card:validityChange', function (event) {
+ *     // the card form went from invalid to valid or valid to invalid
+ *   });
+ * });
  */
 
 /**
@@ -252,6 +298,54 @@ HAS_RAW_PAYMENT_DATA[constants.paymentMethodTypes.applePay] = true;
  * This event is emitted when the customer selects a new payment option type (e.g. PayPal, PayPal Credit, credit card). This event is not emitted when the user changes between existing saved payment methods. Only relevant when accepting multiple payment options.
  * @event Dropin#paymentOptionSelected
  * @type {Dropin~paymentOptionSelectedPayload}
+ */
+
+/**
+ * The underlying [hosted fields `binAvailable` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:binAvailable).
+ * @event Dropin#card:binAvailable
+ * @type {Dropin~card:binAvailable}
+ */
+
+/**
+ * The underlying [hosted fields `blur` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:blur).
+ * @event Dropin#card:blur
+ * @type {Dropin~card:blur}
+ */
+
+/**
+ * The underlying [hosted fields `cardTypeChange` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:cardTypeChange).
+ * @event Dropin#card:cardTypeChange
+ * @type {Dropin~card:cardTypeChange}
+ */
+
+/**
+ * The underlying [hosted fields `empty` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:empty).
+ * @event Dropin#card:empty
+ * @type {Dropin~card:empty}
+ */
+
+/**
+ * The underlying [hosted fields `focus` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:focus).
+ * @event Dropin#card:focus
+ * @type {Dropin~card:focus}
+ */
+
+/**
+ * The underlying [hosted fields `inputSubmitRequest` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:inputSubmitRequest).
+ * @event Dropin#card:inputSubmitRequest
+ * @type {Dropin~card:inputSubmitRequest}
+ */
+
+/**
+ * The underlying [hosted fields `notEmpty` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:notEmpty).
+ * @event Dropin#card:notEmpty
+ * @type {Dropin~card:notEmpty}
+ */
+
+/**
+ * The underlying [hosted fields `validityChange` event](http://braintree.github.io/braintree-web/{@pkg bt-web-version}/HostedFields.html#event:validityChange).
+ * @event Dropin#card:validityChange
+ * @type {Dropin~card:validityChange}
  */
 
 /**
@@ -375,16 +469,10 @@ Dropin.prototype._initialize = function (callback) {
       }
     });
 
-    self._model.on('paymentMethodRequestable', function (event) {
-      self._emit('paymentMethodRequestable', event);
-    });
-
-    self._model.on('noPaymentMethodRequestable', function () {
-      self._emit('noPaymentMethodRequestable');
-    });
-
-    self._model.on('paymentOptionSelected', function (event) {
-      self._emit('paymentOptionSelected', event);
+    PASS_THROUGH_EVENTS.forEach(function (eventName) {
+      self._model.on(eventName, function (event) {
+        self._emit(eventName, event);
+      });
     });
 
     return self._setUpDependenciesAndViews();