+++ src/translations/en_AU.js
@@ -20,6 +20,7 @@ module.exports = {
   "fieldTooLongForCardholderName": "Cardholder name must be less than 256 characters.",
   "fieldEmptyForNumber": "Please fill out a number.",
   "fieldEmptyForPostalCode": "Please fill out a postcode.",
+  "fieldInvalidForCardholderName": "This cardholder name isn't valid.",
   "fieldInvalidForCvv": "This security code is not valid.",
   "fieldInvalidForExpirationDate": "This expiry date is not valid.",
   "fieldInvalidForNumber": "This card number is not valid.",