+++ test/integration/request-payment-method.test.js
@@ -1,5 +1,3 @@
-require('./helper');
-
 describe('Drop-in#requestPaymentMethod', function () {
   beforeEach(function () {
     browser.reloadSessionOnRetry(this.currentTest);