+++ test/unit/views/payment-sheet-views/base-paypal-view.js
@@ -143,7 +143,10 @@ describe('BasePayPalView', () => {
     it('calls paypal.Button.render', () => {
       return testContext.view.initialize().then(() => {
         expect(testContext.paypal.Button.render).toBeCalledTimes(1);
-        expect(testContext.paypal.Button.render).toBeCalledWith(expect.any(Object), '[data-braintree-id="paypal-button"]');
+        expect(testContext.paypal.Button.render).toBeCalledWith(
+          expect.any(Object),
+          expect.stringMatching(/#braintree--dropin__.*\[data-braintree-id="paypal-button"\]/)
+        );
       });
     });
 
@@ -757,7 +760,10 @@ describe('BasePayPalView', () => {
 
       it('uses the PayPal button selector', () => {
         return testContext.view.initialize().then(() => {
-          expect(testContext.paypal.Button.render).toBeCalledWith(expect.any(Object), '[data-braintree-id="paypal-button"]');
+          expect(testContext.paypal.Button.render).toBeCalledWith(
+            expect.any(Object),
+            expect.stringMatching(/#braintree--dropin__.*\[data-braintree-id="paypal-button"\]/)
+          );
         });
       });
     });
@@ -820,7 +826,10 @@ describe('BasePayPalView', () => {
         testContext.view._isPayPalCredit = true;
 
         return testContext.view.initialize().then(() => {
-          expect(testContext.paypal.Button.render).toBeCalledWith(expect.any(Object), '[data-braintree-id="paypal-credit-button"]');
+          expect(testContext.paypal.Button.render).toBeCalledWith(
+            expect.any(Object),
+            expect.stringMatching(/#braintree--dropin__.*\[data-braintree-id="paypal-credit-button"\]/)
+          );
         });
       });